Relocate master_list_line repository test to repository tests
