+++ repository/src/db_diesel/master_list_line.rs
@@ -74,41 +74,3 @@ fn create_filtered_query(
 
     Ok(query)
 }
-
-#[cfg(test)]
-mod test {
-    use domain::{master_list_line::MasterListLineFilter, EqualFilter};
-
-    use crate::{mock::MockDataInserts, test_db, MasterListLineRepository};
-
-    #[actix_rt::test]
-    async fn test_master_list_line_repository_filter() {
-        let (mock_data, connection, _, _) = test_db::setup_all(
-            "test_master_list_line_repository_filter",
-            MockDataInserts::all(),
-        )
-        .await;
-
-        let repo = MasterListLineRepository::new(&connection);
-
-        // Test filter by master_list_id
-        let lines = repo
-            .query_by_filter(
-                MasterListLineFilter::new().master_list_id(EqualFilter::equal_any(vec![
-                    "master_list_master_list_line_filter_test".to_string(),
-                ])),
-            )
-            .unwrap();
-
-        for (count, line) in mock_data
-            .full_master_list
-            .get("master_list_master_list_line_filter_test")
-            .unwrap()
-            .lines
-            .iter()
-            .enumerate()
-        {
-            assert_eq!(lines[count].id, line.id)
-        }
-    }
-}
